requirements.txt
amqp==2.5.2
asgiref==3.2.5
astroid==2.3.3
billiard==3.6.3.0
celery==4.4.2
colorama==0.4.3
Django==3.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-datatables-view==1.19.1
djangorestframework==3.11.0
importlib-metadata==1.6.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.0
vine==1.3.0
zipp==3.1.0
